[REF] Module name from pos_crm to pos_ask_vat
