+++ pos_ask_vat/static/src/js/Popups/TaxIdPopup.js
@@ -1,0 +1,81 @@
+odoo.define("pos_ask_vat.TaxIdPopup", function (require) {
+    "use strict";
+    var core = require("web.core");
+    var _t = core._t;
+
+    const {useState} = owl;
+    const AbstractAwaitablePopup = require("point_of_sale.AbstractAwaitablePopup");
+    const NumberBuffer = require("point_of_sale.NumberBuffer");
+    const {useListener} = require("web.custom_hooks");
+    const Registries = require("point_of_sale.Registries");
+
+    class TaxIdPopup extends AbstractAwaitablePopup {
+        /**
+         * @param {Object} props
+         * @param {Boolean} props.isPassword Show password popup.
+         * @param {Number|null} props.startingValue Starting value of the popup.
+         *
+         * Resolve to { confirmed, payload } when used with showPopup method.
+         * @confirmed {Boolean}
+         * @payload {String}
+         */
+        constructor() {
+            super(...arguments);
+            useListener("accept-input", this.confirm);
+            useListener("close-this-popup", this.cancel);
+            let startingBuffer = "";
+            if (
+                typeof this.props.startingValue === "number" &&
+                this.props.startingValue > 0
+            ) {
+                startingBuffer = this.props.startingValue
+                    .toString()
+                    .replace(".", this.decimalSeparator);
+            }
+            this.state = useState({buffer: startingBuffer});
+            NumberBuffer.use({
+                nonKeyboardInputEvent: "numpad-click-input",
+                triggerAtEnter: "accept-input",
+                triggerAtEscape: "close-this-popup",
+                state: this.state,
+            });
+        }
+        get decimalSeparator() {
+            return this.env._t.database.parameters.decimal_point;
+        }
+        get inputBuffer() {
+            if (this.state.buffer === null) {
+                return "";
+            }
+            if (this.props.isPassword) {
+                return this.state.buffer.replace(/./g, "•");
+            }
+            return this.state.buffer;
+        }
+        confirm(event) {
+            const bufferState = event.detail;
+            if (bufferState.buffer !== "") {
+                super.confirm();
+            }
+        }
+        sendInput(key) {
+            this.trigger("numpad-click-input", {key});
+        }
+        getPayload() {
+            return NumberBuffer.get();
+        }
+    }
+    TaxIdPopup.template = "TaxIdPopup";
+    TaxIdPopup.defaultProps = {
+        confirmText: _t("Ok"),
+        cancelText: _t("Cancel"),
+        title: _t("Confirm ?"),
+        body: "",
+        startingValue: null,
+        isPassword: false,
+    };
+
+    Registries.Component.add(TaxIdPopup);
+
+    return TaxIdPopup;
+});
