+++ pos_ask_vat/views/pos_config.xml
@@ -1,0 +1,56 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo>
+    <record model="ir.ui.view" id="view_pos_config_form">
+        <field name="name">view.pos.config.form</field>
+        <field name="model">pos.config</field>
+        <field name="inherit_id" ref="point_of_sale.pos_config_view_form" />
+        <field name="arch" type="xml">
+            <div id="inventory_location" position="after">
+                <h2>Customers</h2>
+                <div class="row mt16 o_settings_container" id="customers">
+                    <div class="col-12 col-lg-6 o_setting_box">
+                        <div class="o_setting_left_pane">
+                            <field name="iface_pos_ask_vat" id="iface_pos_ask_vat" />
+                        </div>
+                        <div class="o_setting_right_pane">
+                            <label for="iface_pos_ask_vat" />
+                            <div class="text-muted">
+                                Create customers in POS workflow
+                            </div>
+                            <div
+                                class="content-group mt16"
+                                attrs="{'invisible': [('iface_pos_ask_vat', '=', False)]}"
+                            >
+                                <div class="row mt16">
+                                    <label
+                                        string="VAT Question"
+                                        for="pos_ask_vat_question"
+                                        class="col-lg-3 o_light_label"
+                                    />
+                                    <field
+                                        name="pos_ask_vat_question"
+                                        id="pos_ask_vat_question"
+                                    />
+                                </div>
+                                <div
+                                    class="row mt16"
+                                    attrs="{'invisible': [('pos_ask_vat_question', 'not in', ['payment', 'order'])]}"
+                                >
+                                    <label
+                                        string="Autocreate Partner"
+                                        for="pos_ask_vat_auto_create_partner"
+                                        class="col-lg-3 o_light_label"
+                                    />
+                                    <field
+                                        name="pos_ask_vat_auto_create_partner"
+                                        id="pos_ask_vat_auto_create_partner"
+                                    />
+                                </div>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </field>
+    </record>
+</odoo>
