--- pos_crm/views/pos_order.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<!-- Copyright 2023 KMEE
-     License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl). -->
-<odoo>
-
-    <record model="ir.ui.view" id="pos_order_form_view">
-        <field name="name">pos.order.form (in pos_crm)</field>
-        <field name="model">pos.order</field>
-        <field name="inherit_id" ref="point_of_sale.view_pos_pos_form" />
-        <field name="arch" type="xml">
-            <field name="partner_id" position="after">
-                <field name="customer_tax_id" />
-            </field>
-        </field>
-    </record>
-
-
-
-</odoo>
