--- pos_crm/views/pos_template.xml
@@ -1,37 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<odoo>
-    <data>
-        <template
-            id="pos_crm_assets_backend"
-            name="pos_crm_assets_backend"
-            inherit_id="point_of_sale.assets"
-        >
-
-            <xpath expr="." position="inside">
-                <script type="text/javascript" src="/pos_crm/static/src/js/db.js" />
-                <script type="text/javascript" src="/pos_crm/static/src/js/models.js" />
-                <script
-                    type="text/javascript"
-                    src="/pos_crm/static/src/js/PaymentScreen.js"
-                />
-                <script
-                    type="text/javascript"
-                    src="/pos_crm/static/src/js/Popups/TaxIdPopup.js"
-                />
-                <script
-                    type="text/javascript"
-                    src="/pos_crm/static/src/js/Screens/OrderManagementControlPanel.js"
-                />
-                <script
-                    type="text/javascript"
-                    src="/pos_crm/static/src/js/Screens/OrderFetcher.js"
-                />
-                <script
-                    type="text/javascript"
-                    src="/pos_crm/static/src/js/Misc/NumberBuffer.js"
-                />
-            </xpath>
-        </template>
-
-    </data>
-</odoo>
