+++ setup/pos_ask_vat/odoo/addons/pos_ask_vat
@@ -1,0 +1,1 @@
+../../../../pos_ask_vat
