--- setup/pos_crm/odoo/addons/pos_crm
@@ -1,1 +1,0 @@
-../../../../pos_crm
